# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BotTalk API
  description: |
    BotTalk 是一个开源的微信推送服务（类似 Server酱），通过微信 iLink Bot 向用户发送消息推送。

    ## 认证方式

    - **SendKey**：用于推送 API，32 位十六进制字符串，支持路径传参、Query 传参和 Bearer Token 三种方式。
    - **Session Cookie**：用于管理 API，通过登录或扫码注册后自动设置，有效期 30 天。

    ## 响应格式

    推送 API 使用 `code` + `message` 风格；管理 API 使用 `success` 布尔值风格。
  version: 1.0.0
  contact:
    name: BotTalk
    url: https://bot-talk.com
servers:
  - url: https://bot-talk.com
    description: 托管服务
  - url: http://localhost:3000
    description: 本地开发 / 自部署
...
components:
  securitySchemes:
    SendKeyQuery:
      type: apiKey
      in: query
      name: key
      description: 通过 query 参数传递 SendKey
    SendKeyBearer:
      type: http
      scheme: bearer
      description: 通过 Authorization Bearer 头传递 SendKey
    SessionCookie:
      type: apiKey
      in: cookie
      name: connect.sid
      description: 登录 session cookie
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
    UnauthorizedError:
      type: object
      properties:
        error:
          type: string
          example: "Unauthorized. Please log in."
    InternalError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Internal error"

    # ── 数据模型 ──
...
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      summary: 邮箱注册
      description: 通过邮箱注册新用户，自动生成 SendKey 并建立登录 session。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                  description: 邮箱地址
      responses:
        '200':
          description: 注册成功或业务错误
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      success:
                        type: boolean
                        example: true
                      data:
                        type: object
                        properties:
                          id:
                            type: integer
                          email:
                            type: string
                          send_key:
                            type: string
                  - $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalError'
  /api/auth/login:
    post:
      tags: [Auth]
      summary: SendKey 登录
      description: 通过 SendKey 登录，建立 session。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [send_key]
              properties:
                send_key:
                  type: string
                  description: 用户的 SendKey
      responses:
        '200':
          description: 登录成功或业务错误
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      success:
                        type: boolean
                        example: true
                      data:
                        type: object
                        properties:
                          id:
                            type: integer
                          email:
                            type: string
                          send_key:
                            type: string
                          channel_count:
                            type: integer
                  - $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalError'
  /api/auth/me:
    get:
      tags: [Auth]
      summary: 获取当前用户信息
      description: 获取当前登录用户信息，需要登录 session。
      security:
        - SessionCookie: []
      responses:
        '200':
          description: 用户信息
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                      email:
                        type: string
                      send_key:
                        type: string
                      role:
                        type: string
                        enum: [user, admin]
                      channel_count:
                        type: integer
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalError'
  /api/auth/logout:
    post:
      tags: [Auth]
      summary: 退出登录
      description: 退出登录，销毁 session。
      security:
        - SessionCookie: []
      responses:
        '200':
          description: 登出成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalError'

  # ════════════════════════════════════════════════════════════════════
  #  Key API
  # ════════════════════════════════════════════════════════════════════
...